moved clienthandler to server.java
