--- Multithread-ISS/tempCodeRunnerFile.java
@@ -1,15 +1,0 @@
-removeClientHandler();
-        try {
-            if(bufferedReader != null) {
-                bufferedReader.close();
-            }
-            if(bufferedWriter != null) {
-                bufferedWriter.close();
-            }
-            if(socket != null) {
-                socket.close();
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
